Ensure removed probes are well removed from s3Pools map
Currently when a service must be removed we well stop the probe but doesn't remove the service from the
s3Pools map.
This map is used to know which service is monitor. If for external reason thee service is a bit flaky in consul
the probe will be well removed and at thee next WatchPools check the service will still be seen as monitor by the probe.
So if it is again available in consul we do not recreate the probe leading to not monitored service.

Also during probe creation if for any reason we failed to prepare the probes (prepareLatencyBucket or prepareDurabilityBucket), the
probe is not well started and  never try to be started anymore